rename to `getSilentParts()`
