+++ packages/renderer/rust/commands/mod.rs
@@ -3,7 +3,7 @@ use crate::errors::ErrorWithBacktrace;
 use crate::image::{save_as_jpeg, save_as_png};
 use crate::opened_video_manager::OpenedVideoManager;
 use crate::payloads::payloads::CliInputCommandPayload;
-use crate::{ffmpeg, get_silences};
+use crate::{ffmpeg, get_silent_parts};
 use std::io::ErrorKind;
 
 pub fn execute_command(opts: CliInputCommandPayload) -> Result<Vec<u8>, ErrorWithBacktrace> {
@@ -55,7 +55,7 @@ pub fn execute_command(opts: CliInputCommandPayload) -> Result<Vec<u8>, ErrorWit
                 "silencedetect=n={}dB:d={}",
                 _command.noiseThresholdInDecibel, _command.minDuration
             );
-            let res = get_silences::get_silences(_command.src, filter.to_string())?;
+            let res = get_silent_parts::get_silences(_command.src, filter.to_string())?;
             let str = serde_json::to_string(&res)?;
             Ok(str.as_bytes().to_vec())
         }
