+++ packages/renderer/rust/get_silent_parts.rs
