+++ packages/renderer/rust/main.rs
@@ -3,7 +3,7 @@ mod compositor;
 mod errors;
 mod ffmpeg;
 mod frame_cache;
-mod get_silences;
+mod get_silent_parts;
 mod global_printer;
 mod image;
 mod logger;
